requirements.txt
flask
dashscope
manim
python-dotenv
gunicorn
jieba
scikit-learn
